requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
prisma==0.11.0
pymupdf==1.23.6
python-docx==1.1.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
